+++ components/work/projects-section.js
@@ -12,7 +12,7 @@ export default function ProjectsSection({ projects }) {
             subtitle={project.subtitle}
             coverAlt={project.coverImage.alt}
             coverSrc={urlFor(project.coverImage).width(640).height(640).url()}
-            href="#"
+            href={`/projects/${project.slug}`}
           />
         );
       })}