+++ studio/resolveProductionUrl.js
@@ -4,5 +4,5 @@ const previewSecret = process.env.SANITY_STUDIO_PREVIEW_SECRET;
 const projectUrl = constants.siteUrl;
 
 export default function resolveProductionUrl(document) {
-  return `${projectUrl}/api/preview?secret=${previewSecret}&slug=${document.slug.current}`;
+  return `${projectUrl}/api/preview-${document._type}?secret=${previewSecret}&slug=${document.slug.current}`;
 }